nanostack: allow to forward realm-local multicast
The IPv6 stack strictly respect zones for each interface. Multicast
frames are never forwarded if the zones does not match.

By default, each interface has its own zones for scopes ff01::
(interface-local), ff02:: (link-local), ff03:: (realm-local).

Thus, it is not possible to forward theses multicast frames between
interface. This makes sense if we consider that backhaul is an ethernet
interface. However, in wsbrd, from user perspective, the tun interface
is a RF interface. So it should be possible to send these multicast
frames on the network.

So, this patch also modify multicast_fwd_full_for_scope() to also fix the
zone.