os: fix events handling with high workloads
We use a fifo to notify events to the main loop.

Currently, the main loop pulls only one event at time. For each event,
several new events can pushed in the pipe. At the end, the internal fifo
buffer can be full. Then, it is no more possible to append more data and
call to write() never return.

This patch makes write to fifo non-blocking. It generate no limitation
since it not necessary to stack the events.

We can go even further. The internal fifo buffer only need to be big
enough for one event. In order to avoid any race condition, we set its
size to two events (even if I fill it is an unneeded precaution).

Since the initialisation of the fifo starts to be long, we add function
for that.