nanostack: fix overflow in IE header search
The function mac_ie_header_sub_id_discover searches for a specific
Information Element (IE) header in an array, given its ID.  Every header
starts with 16 bits which give the length of the header (7 bits) and its
ID (8 bits) as extracted by the function mac_ie_header_parse.

The termination condition for the search loop is based on a decrementing
variable (length) supposed to represent the length of the input array
that is left to explore.

However, since length is unsigned, and the header length is not checked
to be coherent with the total length of the input array, length can
underflow and thus increment instead of decrement.

Finally since every loop execution the pointer to read is incremented,
so it can result in reading out of the allocated memory, causing a
segmentation fault.