wsbrd: enable DHCPv6 relay agent on tun and rcp interfaces
both are needed in order to relay DHCPv6 messages:
  - The solicit requests from the rank 1 nodes (closest neighbors) need
    the relay running on the rcp interface.
  - The solicit requests from nodes of rank 2 or more sends their
    requests to the relay running on the tun interface.

We have to init and configure a DHCPv6 client instance on the tun
interface after bringing it's been brought up because the dhcp relay
can't be enabled without one.

We also ignore ipv6_prefix when dhcp_server is also specified.
In accordance with the documented behavior in the configuration file.

[Jérôme: reword comment in wsbrd.conf]