wsbrd: fixed dropped message on the RCP with retransmission to it
An RX buffer overflow sometimes occurs on the RCP, causing bytes to be
lost. It happens more frequently when aggregated packets are sent to it.
The fix is to keep a number (64 here) of sent packets from the host to
the RCP and hoping the CRC of the packet hasn't been lost, otherwise
we may run into problems such as :

- the border router cannot sent data to a node anymore (WI_SUN-1276)