stack: only use randlib with mbedtls
mbedtls uses two entropy sources: default and randlib. Both use the
syscall getrandom, but the default source does it using
'syscall(SYS_getrandom, ...)' instead of the C wrapper 'getrandom'.
This makes it hard to wrap for capture/replay and fuzzing  because
'syscall' is a variadic function.

The default source is not designed to be disabled so it must be done
through the internal API of mbedtls, which changed after v3.0.0.
mbedtls will only work if there is at least one strong entropy source.