wsbrd: introduce a wsbrd API version
Until now, only the RCP had an API version. However, we plan to change
the SPINEL_STREAM_RAW API. For this change, the RCP has no way to know
if the wsbrd support the old or the new version of the API. So we
introduce a version to wsbrd.

The rules to increment the APIs are now:
  - increment the API of the RCP each time a PROP_SET is changed
  - increment the API of the wsbrd each time a PROP_IS is changed