wsbrd: tun0 IPv6 addresses are now derived from the hwaddr of the RCP
WI_SUN-1684: used libnl and the hwaddr of the RCP to set the addresses
of tun0 (ll and gua).
WI_SUN-1686: the route to the Wi-SUN network for both ll and gua addr
are added automatically when the interface is brought up.
Another link local address is created by libnl when the interface is
brought up, we remove it right away.