stack: DHCP server now uses native socket
We are dropping the internal relay, it's only kept as is for wsnode.

As a result we are also deprecating the 'dhcpv6_server' option in
wsbrd.conf. As a replacement we are adding a new option 'internal_dhcp'
in wsbrd.conf.

If 'internal_dhcp' is false, the internal dhcp server does not start.
Thus allowing another dhcp server or relay to run on the same host such
as ISC dhcpd or dhcrelay.

Down the line this could allow us to drop all the dhcp code from wsbrd.

See also WI_SUN-1575.

[Jérôme:
  - Fix use of msghdr with wsnode.
  - Change the commit log to mention 'internal_dhcp' instead of
    'external_dhcp' ]