stack: update neighbor_timings on rcp when removing temp entries
ns_fhss_ws_drop_neighbor() calls have been added to remove these entries
from the neighbor_timings array on the rcp. This patch fixes a leak on
the rcp.

this patch doesn't need to be ported to the EFR as it doesn't use the
same mechanism to manage its neighbors.

see WI_SUN-803