stack: fix missing call to ns_fhss_ws_drop_neighbor for temp multicast
This code is responsible for replacing the last entry in the multicast
neighbors temporary list by a new one. This new entry is placed
at the start of the temporary multicast neighbor list.
This list seems useless as it is updated when we receive a DIS from a
neighboring node. The nodes address the router with unicast during join
state 4. The multicast neighbors temporary list is useless on the BR as
of now.

For the sake of completeness a patch would be necessary on the rcp side
in case the neighbor being dropped in the fhss message was already
dropped in a previous fhss message.