wbsrd: join multicast groups ff02::1a, ff03::1, ff03::2 and ff03::fc
Linux automatically subscribes to ff02::1 and ff02::2, we subscribe to
the rest to stay in line with the Wi-SUN spec and the behavior of the
stack.

Should we add a Dbus API to join multicast groups later, we will have to
both add the group with the stack API and subscribe to the mcast group
on the Linux host side.