dbus: get rid of trivial int setter
From sd-bus documentation (man sd_bus_add_object):
The setter and getter methods may be omitted (specified as NULL), if the
property is one of the basic types or "as" in case of read-only properties.

This does not work for dbus_get_string because userdata is a (char *)
instead of a (char **).