stack: fix concurrent authentications
The struct sec_prot_gtk_keys_t is shared between all the supplicants.
Until now, we stored data associated to a supplicant into this struct.
So, things started to get wrong as soon as several nodes tried to
authenticate.

However, we still need a struct to store data duplicated in LGTK and
GTK. We introduce sec_prot_gtk_t for that purpose.