stack: read new network-keys format
Note this patch slightly change the way the "updated" flag is set. This
flags ask to the stack to write the new keys to the NVM. Before this
change, it was set to false to save one write access. However, now if
the use only set the first, the second one can be retrieved from the
NVM, so we have to be conservative on this flags. So finally, we just
don't change it.

See also WI_SUN-1322