conf: rename line_no in linenr
This patch is only here to align commandline with key_value_storage.

Some may wonder if "linenr" is better than "lineno". According this
post[1], "lineno" is preferred. It also seems that a majority of
developers use lineno:
    linux-kernel$ git grep 'line.\{0,3\}nr' | wc -l
    381
    linux-kernel$ git grep 'line.\{0,3\}no' | wc -l
    1235

However, I find odd when an abbreviation includes one or more letters
that do not appear in the word being abbreviated.

[1]: https://english.stackexchange.com/q/43095