stack: drop delay before RPL start
In the old days, wsbrd get its IP address from a DHCP server. RPL
service couldn't start before to have an IP configuration. So there was
a system to delay the start of the RPl service.

Since the network configuration is now fully available on start of
wsbrd, this delay does not make sense anymore.

Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>