wsbrd: remove fw_api_older_than()
Declaring it in wsbr.h forced the inclusion of version.h, which includes
bits.h, which defines a function named bitand(). However 'bitand' is a C++
keyword (alias for bit operator '&'), instead of a macro provided by
iso646.h as in C... Nonetheless this function does not bring much
compared to directly calling version_older_than().