utils: fix dhcp relay forward handling
If a malformed relay packet arrived with less bytes than intended,
computing the length of the dhcp options would result in an underflow,
because 'req.data_size - 34' is passed as a unsigned value (size_t) to
dhcp_handle_request().

Using a signed integer would have worked, but I opted for proper parsing
of the DHCP header instead.

This was found by AFL++ using wsbrd-fuzz with the packet '01 00 0c'
which resulted in a segmentation fault. Since the minimized testcase is
small, I'll paste it fully here in text (does not include RCP init):
30 81 80 7c 01 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 01 00 0c 30 30 7e