wsbrd: fix store_rf_config_list()
- The base PHY was not added
- A single MCS was picked per OFDM PHY mode ID
- Only the 1st entry from a multiphy could be picked as base PHY

Note that since we are not retrieving the regulatory domain from the
radioconf, the first entry that match our base PHY will be selected.
For example NA and MX share some channel configuration, but only NA can
do mode switch. So if NA comes first in our radioconf, but we select an
MX base PHY, we will end up advertizing mode switch as if we had
selected the corresponding NA PHY.

See also WI_SUN-2318.