wsbrd: introduce RCP_HAS_RF_CONFIGURATION
This fixes an issue with the way unsupported radio configurations are
handled. The RCP is supposed to reset when it receives an invalid PHY
configuration. However in some cases it would reset before finishing to
send its confirmation (PROP_IS/WS_RF_CONFIGURATION), leading to the
error message:
    MAC layer has been reset. Operation not supported
instead of:
    RF configuration not supported by the RCP

This may break wsbrd-fuzz since some commands are refused while
RCP_INIT_DONE is not set.

See also WI_SUN-2318.