tun: force MPL tunnelling when dst is ff03::fc
MPL specifies that if the destination address is those of the a MPL
domain, the packet does not need to be tunnelled, but the MPL option
needs to be added. I could not find an easy way to add the MPL option
since the existing code is designed to build entirely the IPv6 packet
header and fill the options in a sort of implicit state machine way.
But we insert a fully built packet received from the TUN interface,
which makes it difficult to reuse existing code to fill the MPL option.
The crappy solution proposed here is to tunnel the packet as if the
destination address was any multicast address.

WI_SUN-2332