stack: fix channel exclusion advertisement
Bit ordering is awkward in the Wi-SUN specification since figures use
'on the air' ordering (lsb first), which is reversed from the 'in
memory' ordering. This is described in Wi-SUN 1.1v06 6.3.2.1.1.

Note that only advertised masks are affected, effective masks were
correct, which would lead to interoperability issues. This change is
also welcome as it eliminates the need to reverse masks.

The same bug was present in the Silabs embedded Wi-SUN stack, thus
interoperability with versions older than 23q2-patch2 (GSDK 4.3.2 or
Wi-SUN 1.7.1) is compromised.