fuzz: support new RPL implementation
The new RPL implementation uses recvmsg() which was not previously
supported. Moreover, the destination address is retrieved using
ancillary data so it must now be stored in the capture file.

Note that this breaks the format of capture files, but it does not
really matter since these are only meant to be replayed with the
specific wsbrd version they were captured with.