tbu: reuse main config file in wstbu-dhcpv6-relay
The initial intended behavior was to copy the dhcpv6_server line from
the main TBU config file to another one specifically for the DHCPv6
relay. However due to an oversight, this was never done from the TBU
server script. In any case, systemd is able to parse .ini files which
makes the initial idea useless.