stack: introduce rpl_root.compat
Path Sequence handling appears to cause interoperability issues, which
is understandable as the specification is unclear with regards to some
aspects. Some routers have been observed to:
- Advertise new information (eg. new parents) without incrementing their
  path sequence.
- Advertise information using an out-of-date path sequence.

Additionnally, issues have been observed concerning the handling of the
source routing header.

This flag allows to degrade the RPL implementation in favor of a more
simple behavior, but hopefully more functional in the context of
networks including potentially faulty RPL implementations.