stack: remove network_name from pae_auth_t
network_name was only assigned in ws_pae_auth_nw_info_set().

The check in the same function to know whether the install of (l)gtks
should be forced or not was also useless as this function is in fact not
called twice but only once in the call chain.

Therefore, this commit also drops the force_install parameter of
ws_pae_auth_network_keys_from_gtks_set() and subsequent functions.