stack: assume key is live when no liveness KDE present
When supplicants issue an EAPOL Key Request packet, they are supposed
to include a liveness KDE to tell the authenticator which keys they
possess. If the KDE for a specific key type (GTKL or LGTKL) is not
included, the authenticator now assumes that the supplicant does not
want to receive any key of the relevant type (which is equivalent to
assuming that the supplicant possesses all keys of that type).

WI_SUN-2873