ws: stop removing temporary mac neighbor after tx confirm
It appears that we were deleting temporary mac/ipv6 neighbor after using each
entry briefly in join state 4.
We believe this is legacy and was done to save memory.
Doing so today is not as relevant as it was. Leaving the entry in the
cache and letting the cache manager handle the deletion makes more
sense.