utils: make CRC more generic
This allows to implement all variants of CRC-16 with poly=0x1021,
refin=true and refout=true, and also to compute the CRC incrementally
in case of a split buffer. Users should pass the init value as the crc
argument when first calling crc16(), and manually perform a xor on the
output when necessary.

Reviewed-by: Jérôme Pouiller <jerome.pouiller@silabs.com>