ws: improve ws_neighbor_calculate_ufsi_drift() trace
Up to now, the calculated drift was not based on floating point values
which prevented from providing an accurate UFSI drift to the user.

In addition to fixing this issue, this patch also modifies what
information is traced and when such information is traced.
We chose to trace this information only if the drift calculus is
performed on a window at least equal to 10s to ensure 0.1ppm of
procision.

Note that this solution is best effort at the time and can be improved
by keeping a record of the ufsi over time.