ws: drop MAC level NUD (neighbor unreachability detection)
ws_bootstrap_neighbor_nud_notify() was used to start a "NUD" process when
a neighbor was considered "unreachable".
This process consisted in sending NS to the neighbor and expecting it to
ACK such NS if it was still "reachable". After some retries, if the neighbor
was not answering these NS, it would simply get deleted.

This mechanism is NOT a MAC level one and is NOT standard.
In addition, it got obselete in commit 6ca7838836f2 ("stack: drop legacy
socket API") in which a call to ack_receive_cb, allowing to remove the
neighbor from the active_nud_process list, was removed.
Since then, a neighbor would systematically be removed once a NUD
process started, whether an ACK was received or not.