rpl: accept RPI with SenderRank = 0
From RFC 6550 section 11.2:

   SenderRank: 16-bit field set to zero by the source and to
         DAGRank(rank) by a router that forwards inside the RPL network.

The reasoning behing setting SenderRank to 0 is unclear, and it was
confirmed to me by Pascal Thubert that the goal is to allow RUL devices
(equivalent of LFN) to insert a RPI option without too much effort,
allowing their parent to not perform tunnelling.

However it does not really make sense to set the SenderRank to 0 from
the perspective of a fully functional RPL node. However since it was
written this way in the specification, some routers do it.