sec: introduce hmac_md_tlv_*()
Until now, hmac_md_calc() had an "interesting feature". When the payload
was in TLV format it was able to read the Length from the payload and
shrink the data_size if needed. When the payload was not a TLV, the
length read was greater than the data_size, so data_size was not
shrunk.

This behavior was highly dangerous. This patch introduce hmac_md_tlv_*()
which read the payload length explicitly.

Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>