ws: refresh and trust ws_neighbor if an associated ipv6 neighbor exists
With this patch we ensure a newly created ws_neighbor has a lifetime
that matches its corresponding ipv6 neighbor lifetime.
This is useful only upon reboot when we receive a packet from a device
that was not part of our ws_neighbor_table but whose information were
restored from the ipv6 neighbor cache storage.

Reviewed-by: Jérôme Pouiller <jerome.pouiller@silabs.com>