sec: fix bad memory handling architecture 2/2
The previous commit revealed a bad memory handling architecture in the
security module.

To be more specific, kmp_socket_if_send() is the function used to
transmit a message in this module.
This function was freeing the pdu parameter on success deliberately
without being the owner of this pointer.

In addition, kmp_sec_prot_conn_send(), the caller of
kmp_socket_if_send(), was freeing this pointer only if the function was
failing, which was also a bad memory handling as this function is not
the owner of the pointer either.

These two frees were dropped in the previous commit to fix the
discovered segfault.
This commit fixes the aforementioned architecture issue and ensures the
pdu parameter is freed after usage by the owner of this pointer.

WI_SUN-2903