wsbrd: be more reactive on pcap fifo close
Before, when the FIFO read end was closed, it would not be detected until
there was an actual packet to be written. This meant that if the FIFO
was closed and reopened in an interval where no packet was sent, then no
start block would be written, causing confusion on the read side.