wsbrd: WARN when both pan_id and (L)GTKs are set in wsbrd.conf
Having this combination will cause issues if the user delete wsbrd's
cache and restarts the border router with the same values in wsbrd.conf.
Nodes will act like everything is fine as the pan_id is still the same
and as they are still able to decrypt packets.
wsbrd actually has no RPL info about the nodes and its IPv6 neighbor
cache is empty, which makes the network unusable until the nodes update
their information.