clean: arbitrarily drop buffer_priority_set()
After an internal debate we chose to drop the QoS handling.
This legacy mechanism was not proven very useful during our tests.

We also believe that this should be Linux's job to handle this priority
stuff.

In addition, commit 237620827c ("Activated higher priority by traffic
class CS6 for NS/NA and RPL, EAPOL/ DHCP Relay messages.") (and others)
integrated the usage of this mechanism, but the work done in these commits
got gradualtely dropped as code refactoring was done. As never an issue
linked to not using this mechanism was found, we believe going further
and entirely dropping this mechanism will not cause any problem.