hwping: fix ping requests
APIv2 has an additional field to allow for a different payload size
between the request and the confirmation. This oversight went unnoticed
because the way the payload is contructed starts with 00 01 02..., which
was interpreted as "a 1 byte buffer with contents 0x02". Testing the TX
mode makes the issue apparent as the RCP expects to read payload length
of 0 but instead finds that there are not enough bytes.