ws: do not assert on unexpected retry count in data confirmation
Some RCP versions have a bug where the retry count can exceed the value
requested, or the default value of 20. Additionally, the RCP always
does 20 retries (total on all rates) even if less were requested, and
the filler ones are made using the base PHY. This behavior should
arguably be controlled by the host, but it must be handled since it can
happen.