ws: drop supplicant waiting list limit
If the limit was reached we would remove the oldest supplicant from the
list and add the new one, which was basically like not having a waiting
list. In addition, the content of this list already has an expiration
mechanism which means nothing prevents us from having an infinite
waiting list.

WI_SUN-3336

Reviewed-by: Mathis Marion <mathis.marion@silabs.com>