ws: use ws_neigh in ws_llc_lowpan_mpx_data_request() for assignments
Using TxAckReq was not appropriate considering an ACK is not a
requirement for these assignments.
It is about knowing if DstAddr is a broadcast address or not.
Typically, for EDFE initial frames, an ACK is not requested but it
does not mean enabling mode switch is not possible.

WI_SUN-3258

Reviewed-by: Mathis Marion <mathis.marion@silabs.com>