wsbrd: fix waste of CPU when 6lowpan queue is full
Currently, if lowpan_adaptation_queue_size() returns > 2, the tun
interface is not read. However, poll() still wake up the main loop.
This behavior lead to a useless wake up of the main loop and a waste of
CPU resources.

This issue typically happens during throughput tests when the Linux
stack send frames faster than the RCP is able to send.

See WI_SUN-3494.

Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>