ipv6: do not copy hop limit when creating tunnels
The hop limit selected at a tunnel entry is a local configuration of the
ingress, and has no relation with the received hop limit of the packet
to encapsulate. RFC 2473 section 6.3 describes the use of hop limit in
the case of tunnels in more details.

IP tunnels are normally invisible to tools like traceroute since an
increment of the probe hop limit covers the whole tunnel. However with
this quirk, traceroute needs to increase to hop limit for every
intermediate Wi-SUN router in the tunnel, but it also does not receive
ICMP error messages since they are destined to the ingress.

Here is an example to demonstrate the difference in behavior:

           +  2001:db8::1
 Backhaul--|      | Ethernet link
           +  2001:db8::2           (eth0)
                  | Linux + wsbrd
           +  fd12:3456::200:0:80:3 (tun0)
           |      | Radio
           |  fd12:3456::200:0:80:0
 Wi-SUN----|      | Radio
           |  fd12:3456::200:0:80:1
           |      | Radio
           +  fd12:3456::200:0:80:2

Executing traceroute from the backhaul device:
Before:

    $ traceroute fd12:3456::200:0:80:2
    traceroute to fd12:3456::200:0:80:2 (fd12:3456::200:0:80:2), 30 hops max, 80 byte packets
     1  2001:db8::1 (2001:db8::1)  0.323 ms  0.368 ms  0.280 ms
     2  * * *
     3  * * *
     4  fd12:3456::200:0:80:2 (fd12:3456::200:0:80:2)  40.184 ms  43.054 ms  45.506 ms

After:

    $ traceroute fd12:3456::200:0:80:2
    traceroute to fd12:3456::200:0:80:2 (fd12:3456::200:0:80:2), 30 hops max, 80 byte packets
     1  2001:db8::1 (2001:db8::1)  0.353 ms  0.393 ms  0.318 ms
     2  fd12:3456::200:0:80:2 (fd12:3456::200:0:80:2)  19.228 ms  25.906 ms *

WI_SUN-3547