sec: fix segv on receiving an unexpected frame from EAPOL target
After receiving an EAP reply from a target, this reply is forwarded to
the radius server and the state machine transitions to the
EAP_TLS_STATE_EAP_REQUEST state.
If an incorrect frame is received from the same target while in this
state, we would wrongly try to apply an offset to skip the socket and
EAPOL headers. This would result in a segmentation fault.

WI_SUN-3691