15.4: fix missing src addr in REQ_DATA_TX
The RCP would previously ignore the address field and fill with its
configured address, but this is quite surprising for the user so going
on, the RCP will FATAL() if the requested source address is different
from the one configured. In the future the RCP should accept to transmit
any source address requested, but it is currently unsupported.