rpl: fix timer initialization when loading from cache
When a RPL neighbor is loaded from cache, rpl_storage_load_target()
started a timer with timer_start_rel(). However the timer group has not
yet been initialized since rpl_start() is called afterward. The timer
is removed from the group when calling timer_group_init(), but its
expiration date is not reset. This causes the timer to be in a corrupt
state. Calling `timer_stop()` SegFaults when trying to remove the entry
from the group.