ns3: prevent concurrent g_timer_event access
timer_trig() and timerfd_settime() are not executed by the same thread.
The initial reason for that architecture is that the event is freed
after the callback is called, so it cannot be accessed from
timerfd_settime(). However since it is a reference counted object, we
can simply keep a local reference which can be accessed even after the
event has been processed. Note that Event->Cancel() simply sets a
boolean, it does not perform any real re-scheduling.