wsrd: allow retransmission of 4WH msg 3 and GKH msg 1
If 4WH msg 4 (GKH msg 2) is lost, the authenticator will retransmit 4WH
msg 3 (GKH msg 1). In that case, the supplicant would drop the packet
believing in a key reinstallation attack (KRACK)[1].
This fix allows to ensure the retransmission packet is not dropped and
keeps us safe from KRACK by NOT reinstalling the GTK and consequently
reset the associate frame counter.

[1]: https://www.krackattacks.com