utils: introduce auth_key_accept_frame()
This function introduces the increment of the replay counter as soon as
the MIC is considered valid.
Incrementing the replay counter only when sending a frame would allow an
attacker to replay 4WH msg4.
In the context of Direct Connect, where the TK is used to encrypt
traffic and installed on reception of 4WH msg 4, a replay of the 4WH
msg4 would allow a key reinstallation attack (KRACK)[1].

[1]: https://krackattacks.com/