dhcp: implement relay agent
wsrd requires an implementation of DHCP relay. It would be possible to
rely on an external process. However, it would require a painful setup
for the end user. So, an internal DHCP-relay implementation is needed
for wsrd. Since we need this implementation, there no reason to not use
it in wsbrd. Thus, users will not have to patch, compile and run an
external DHCP relay. In contrary to DHCP servers, relays do not provide
any fancy options. There is no reason to allow the user to use an
alternative DHCP Relay implementation.

WI_SUN-3637