ws: ignore FAN 1.0 entries when selecting mode switch chan mask
This would fatal with "ambiguous channel parameters" in case of channel
configurations with both a FAN 1.0 and a FAN 1.1 definition. Mode switch
should not be performed with FAN 1.0 entries.

This function is now a bit too specific to stay in ws_regdb(). Also
note that none of the other regdb function use FATAL(). BUG() is now
used since this should never happen because the table is statically
defined in the code.