ws: introduce last_dsn_update_tstamp_us
Introducing this new variable was necessary to enable duplicated packet
detection from LFN neighbors.
We cannot rely on LUTT-IE RX timestamp considering the LUTT-IE is only
processed as such if and only if an LUS-IE is also present in a given
frame.
Therefore, the LUTT-IE RX timestamp does not reflect the last RX time of
a frame with a DSN for LFN neighbors.

WI_SUN-4223