rpl: fix parent selection when no parent is found
Previously, if no parent was found during the selection process and
there was no parent at the beginning of it, mrhof would incorrectly
consider the same parent was "kept" while it is actually desirable to
continue with the normal parent selection process.
In addition to fixing this, the parent selection callback is now also
invoked in this case.

WI_SUN-3533