wsrd: improve supplicant failure handling
Instead of immediately emitting a WSRD_EVENT_AUTH_FAIL, we can take a
more measured approach if an RPL parent is available. By denying the
parent and triggering a new parent selection, we allow wsrd to recover
gracefully from this failure and expect the next PC RX to trigger a new
EAPOL Key-Request.
On the opposite, if we were in join state 2, we can deny the neighbor by
setting its routing cost to 0xffff and transition back to join state 1.
Unless the neighbor advertises a new routing cost, we should end up
selecting a different EAPOL target.

WI_SUN-3579