wsrd: remove invalid WSRD_EVENT_RPL_NO_CANDIDATE in join state 4 "choose parent"
Since commit d8b1ec8720 ("rpl: fix parent selection when no parent is
found"), whenever a parent selection is run and no parent is found,
on_pref_parent_change() is called and makes us transition to the
reconnect state. In Join State 4 "choose parent" this is wrong, because
we can easily end up in a situation where a parent selection completes
without finding a suitable candidate, while we did not have any parent
at the beginning of this parent selection.

In addition, this transition was introduced in the early days of the
project and can now be removed since we have the PAN timeout mechanism
acting as a timeout for parent selection.

WI_SUN-3577