wsrd: revoke GTK on hash mismatch in PC frames
Previously, the authenticator was considered the source of truth for
revoking GTKs, as suggested by the specification:

  Wi-SUN FAN 1.1v09 6.3.4.6.3.2.5 FFN Join State 5: Operational
iii. A previously installed GTK[X] is removed from the Border Router
prior to its expiration time. The Border Router disseminated a new PC
Frame, with GTK[X] hash set to 0, into the PAN. Receiving FFNs, still
with local GTK[X] hash value nonzero, detect the hash mismatch and
attempt to acquire the new GTK[X]. The acquisition attempt will reveal
that GTK[X] is no longer valid (via the GTKL KDEs returned by the
authenticator) and the FFN will remove it locally (setting its GTK[X]
hash to 0).

However, the FAN 1.0 Certification Test SEC-GTK-INVALID-1 expects PC
frames to act as the source of truth. Specifically, at the start of the
test, GTK[0], GTK[1], and GTK[2] are installed. Later, GTK[1] and
GTK[2] are revoked, and another key is installed at GTK[1]. This causes
the supplicant to send a Key-Request to the authenticator with its GTKL
indicating GTK[0], GTK[1], and GTK[2] as installed. In this scenario,
the authenticator cannot determine which version of GTK[1] is installed
on the supplicant and considers the key valid.

WI_SUN-4374