auth: use latest slot in new key installation
WI_SUN-4111

  Wi-SUN FAN 1.1v09 6.3.1.1 Configuration Parameters
GTK_EXPIRE_OFFSET: The expiration time of a GTK is calculated as the
expiration time of the GTK most recently installed at the Border Router
plus GTK_EXPIRE_OFFSET.

The authenticator assumed that when a new key is installed, the most
recently installed key is also the active one. With revocation
it may not be the case.

  Wi-SUN FAN 1.1v09 6.5.2.5 Revocation of Node Access
If the remaining lifetime of the currently active L/GTK is less than or
equal to (lifetime / LIFETIME_REDUCTION), the Border Router,
atomically and in specific order, MUST destroy all L/GTKs except the
currently active L/GTK and the next available L/GTK, modify the lifetime
of the next available L/GTK to be (lifetime / LIFETIME_REDUCTION), and
add a new L/GTK (with normal lifetime).