supp: use ptk for mic calculation of 4wh-msg4
In 4-way handshake Message 4, the PTK has been verified and must
be used for MIC calculation. The previous implementation used
the TPTK, which did not cause a bug since they are the same.
However the TPTK could have been removed at this point, especially
when a Secure Element is added.