ipv6: decrement hop limit when forwarding multicast packets to LFN
Note that this code was present in
ipv6_consider_forwarding_unicast_packet(), but not in
ipv6_consider_forwarding_multicast_packet() which refuses to forward
packets outside of MPL. The hop limit is actually decremented by
mpl_forwarder_process_message() for MPL packets. This explains why the
hop limit decrement was initially not included for LFNs.

WI_SUN-4252