rcp: disable CPC reset callback
cpc_init() starts a thread in order to call this function. However this
interfered with the signal masks configured in the main thread,
causing the CPC thread to execute the default signal handler, which for
SIGINT or SIGTERM exits without calling library destructors such as the
one responsible for dumping coverage.

Note that we still end up terminating the process without the callback:

    cpc_read_endpoint: Connection reset by peer

Previously, 2 logs would be printed (supposedly from both threads):

    RCP reset
    cpc_read_endpoint: Connection reset by peer