rpl: do not point route->ip_dest to buf->dst_sa
buf can be resized at any point so it is very unsafe to store pointers
to its fields. In this case, buf->route->ip_dest could get assigned to
buf->dst_sa.address in the case where there is a single hop. The
lifetime of struct rpl_target is guaranteed during the initial
processing of a packet. However we should check if a packet can
remain in a TX queue while the associated RPL target gets deleted.

WI_SUN-4791