wsrd/rpl: prevent parent selection on metric/candidate change until JS 5
This allows wsrd to focus on connecting to the network through the
chosen parent in JS 4.1 before even trying to select a better one.
During the first boot of a network, ETX may fluctuate a lot as all nodes
are trying to join the network at the same time. Allowing parent change
before JS 5 would slow down the network set up.

WI_SUN-4838