wsrd/rpl: send ns(aro) lifetime 1 (60s) instead of 0 when unregistering
A race condition exist: if a packet is sent by the BR through an old
downward route, and a DAO-ACK is sent through a newer (shorter) route,
the DAO-ACK may arrive before the previously sent packet. Sending a
ns(aro) lifetime 0 would prevent that packet from making it to the its
destination as the originator of the ns would be removed from the
routing table of the receiver.

WI_SUN-4702