wsrd/rpl: maintain downward routing until DAO-ACK/disconnection
Sending a ns(aro) lifetime 1 on parent change would potentially mean
removing the downward route after a minute while our new downward route
was never confirmed.
In addition, because we may change parent before a DAO-ACK is received
while a DAO has been sent, we also only unregister from parents that
were not DAO-ACKed after DAO-ACK (or on unregistration). This is
necessary since a DAO that was sent may make it to the BR while the
DAO-ACK may not make it to the originator of the DAO.

WI_SUN-4702