wsrd/supp: always send Key-Requests during Key-Rotation
Making the supplicant fail after 3 retries during key rotation was not
the right approach: we do not want to leave join state 5 as we are most
likely still able to decrypt frames (the rotation window is usually
quite large). However, if we do not leave join state 5 and but the
supplicant still fails, we will not try to fetch the new (L)GTK, and
possibly never will, unless we receive a PC with a new PAN version
during the rotation window.
Therefore, rather than making the supplicant fail, it is better to let
the txalg perform Key-Requests retries, until it eventually gets the new
GTK or the old GTK expires locally.
For the same reason, if a timeout for an answer from the authenticator
is fired, we do not want to leave join state 5, which is why we restart
the Key-Request txalg instead.

WI_SUN-5065